deleted duplicate navbar, invite tab
